api_code/requirements.txt
flask
Flask-RESTful
mysql-connector-python
flask-cors